# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Zhyvo API
  version: 0.1.0
  description: REST API for ephemeral event media rooms. All timestamps are UTC ISO 8601 values.
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: /api/v1
    description: Same-origin API base path
...
paths:
  /auth/anonymous:
    post:
      operationId: createAnonymousSession
      tags: [Auth]
      summary: Create an anonymous identity and device session
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnonymousAuthRequest'
      responses:
        '201':
          description: Session created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '422': {$ref: '#/components/responses/ValidationError'}
        '429': {$ref: '#/components/responses/RateLimited'}
  /auth/telegram:
    post:
      operationId: createTelegramSession
      tags: [Auth]
      summary: Validate Telegram Mini App initData and create a device session
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/TelegramAuthRequest'}
      responses:
        '201':
          description: Telegram identity upserted and session created
          content:
            application/json:
              schema: {$ref: '#/components/schemas/TokenResponse'}
        '401': {description: initData is invalid or expired}
        '429': {$ref: '#/components/responses/RateLimited'}
        '503': {description: Telegram authentication is not configured}
...
  /auth/telegram/oidc:
    post:
      operationId: linkBrowserIdentityToTelegram
      tags: [Auth]
      summary: Exchange an OIDC authorization code and upgrade the current anonymous identity
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/TelegramOIDCRequest'}
      responses:
        '200':
          description: Identity linked; rooms, memberships, uploads and ownership preserved
          content:
            application/json:
              schema: {$ref: '#/components/schemas/TokenResponse'}
        '400': {description: Authorization code, PKCE verifier, nonce or ID token is invalid}
        '409': {description: Current identity cannot be upgraded}
        '503': {description: Telegram browser login is not configured}
  /auth/telegram/link-challenges:
    post:
      operationId: createTelegramBrowserLinkChallenge
      tags: [Auth]
      summary: Create a five-minute single-use browser linking challenge
      responses:
        '201':
          description: Challenge created for the current anonymous identity
          content:
            application/json:
              schema: {$ref: '#/components/schemas/BrowserLinkChallenge'}
        '409': {description: Current identity cannot be linked}
  /auth/telegram/link-challenges/status:
    post:
      operationId: getTelegramBrowserLinkStatus
      tags: [Auth]
      summary: Poll a challenge bound to the current anonymous identity
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/BrowserLinkRequest'}
      responses:
        '200':
          description: Current challenge status
          content:
            application/json:
              schema: {$ref: '#/components/schemas/BrowserLinkChallenge'}
        '404': {description: Challenge not found}
  /auth/telegram/link-challenges/approve:
    post:
      operationId: approveTelegramBrowserLink
      tags: [Auth]
      summary: Approve a browser challenge from an authenticated Telegram Mini App
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/BrowserLinkRequest'}
      responses:
        '200': {description: Challenge approved}
        '409': {description: Challenge expired, denied or already used}
  /auth/telegram/link-challenges/exchange:
    post:
      operationId: exchangeTelegramBrowserLink
      tags: [Auth]
      summary: Consume an approved challenge and upgrade the anonymous browser identity
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/BrowserLinkRequest'}
      responses:
        '200':
          description: Browser identity linked to Telegram
          content:
            application/json:
              schema: {$ref: '#/components/schemas/TokenResponse'}
        '409': {description: Challenge is not approved or is no longer available}
  /auth/refresh:
    post:
      operationId: refreshSession
      tags: [Auth]
      summary: Rotate a refresh token and issue a new access token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token: {type: string, minLength: 32}
      responses:
        '200':
          description: Token rotated
          content:
            application/json:
              schema: {$ref: '#/components/schemas/TokenResponse'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '429': {$ref: '#/components/responses/RateLimited'}
  /auth/me:
    get:
      operationId: getCurrentIdentity
      tags: [Auth]
      summary: Get the current identity
      responses:
        '200':
          description: Current identity
          content:
            application/json:
              schema:
                type: object
                required: [identity]
                properties:
                  identity: {$ref: '#/components/schemas/Identity'}
        '401': {$ref: '#/components/responses/Unauthorized'}
  /auth/session:
    delete:
      operationId: revokeCurrentSession
      tags: [Auth]
      summary: Revoke the current device session
      responses:
        '204': {description: Session revoked}
        '401': {$ref: '#/components/responses/Unauthorized'}
  /rooms:
    get:
      operationId: listMyRooms
      tags: [Rooms]
      summary: List active rooms joined by the current identity
      responses:
        '200':
          description: Active rooms ordered by recent membership activity
          content:
            application/json:
              schema:
                type: object
                required: [rooms]
                properties:
                  rooms:
                    type: array
                    items: {$ref: '#/components/schemas/Room'}
        '401': {$ref: '#/components/responses/Unauthorized'}
    post:
      operationId: createRoom
      tags: [Rooms]
      summary: Create a room
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CreateRoomRequest'}
      responses:
        '201':
          description: Room created, or the original result of an idempotent retry
          content:
            application/json:
              schema:
                type: object
                required: [room, share_path]
                properties:
                  room: {$ref: '#/components/schemas/Room'}
                  share_path: {type: string, example: /r/K7MP2Q}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/ValidationError'}
  /rooms/{slug}/preview:
    get:
      operationId: getRoomPreview
      tags: [Rooms]
      summary: Get public room information before joining
      security: []
      parameters:
        - $ref: '#/components/parameters/RoomSlug'
      responses:
        '200':
          description: Room preview
          content:
            application/json:
              schema: {$ref: '#/components/schemas/RoomPreview'}
        '404': {$ref: '#/components/responses/NotFound'}
        '410': {$ref: '#/components/responses/Gone'}
  /rooms/{slug}/join:
    post:
      operationId: joinRoom
      tags: [Rooms]
      summary: Join a public or protected room
      parameters:
        - $ref: '#/components/parameters/RoomSlug'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                secret: {type: string, description: PIN or password; omit for public rooms}
      responses:
        '200':
          description: Room membership created or refreshed
          content:
            application/json:
              schema:
                type: object
                required: [room]
                properties:
                  room: {$ref: '#/components/schemas/Room'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '410': {$ref: '#/components/responses/Gone'}
        '429': {$ref: '#/components/responses/RateLimited'}
  /invites/{token}/preview:
    get:
      operationId: getManagedInvitePreview
      tags: [Rooms]
      summary: Preview a revocable room invitation
      security: []
      parameters:
        - $ref: '#/components/parameters/InviteToken'
      responses:
        '200':
          description: Invitation and room preview
          content:
            application/json:
              schema: {$ref: '#/components/schemas/RoomInvitePreview'}
        '404': {$ref: '#/components/responses/NotFound'}
        '410': {$ref: '#/components/responses/Gone'}
  /invites/{token}/join:
    post:
      operationId: joinRoomByManagedInvite
      tags: [Rooms]
      summary: Join with the permission carried by a managed invitation
      parameters:
        - $ref: '#/components/parameters/InviteToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                secret: {type: string}
      responses:
        '200':
          description: Room membership created, retained, or upgraded
          content:
            application/json:
              schema:
                type: object
                required: [room]
                properties:
                  room: {$ref: '#/components/schemas/Room'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '410': {$ref: '#/components/responses/Gone'}
  /rooms/{slug}/invites:
    get:
      operationId: listManagedInvites
      tags: [Rooms]
      summary: List managed invitations as the room owner
      parameters:
        - $ref: '#/components/parameters/RoomSlug'
      responses:
        '200':
          description: Invitation management state
          content:
            application/json:
              schema: {$ref: '#/components/schemas/RoomInviteList'}
        '403': {$ref: '#/components/responses/Forbidden'}
    post:
      operationId: createManagedInvite
      tags: [Rooms]
      summary: Create a contributor or view-only invitation as the owner
      parameters:
        - $ref: '#/components/parameters/RoomSlug'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [permission]
              properties:
                permission: {type: string, enum: [contributor, viewer]}
      responses:
        '201':
          description: Invitation created
          content:
            application/json:
              schema: {$ref: '#/components/schemas/RoomInvite'}
        '403': {$ref: '#/components/responses/Forbidden'}
  /rooms/{slug}/share-invite:
    get:
      operationId: getCurrentShareInvite
      tags: [Rooms]
      summary: Get the active contributor invitation as a room member
      parameters:
        - $ref: '#/components/parameters/RoomSlug'
      responses:
        '200':
          description: Current contributor invitation
          content:
            application/json:
              schema: {$ref: '#/components/schemas/RoomInvite'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
  /rooms/{slug}/invites/{token}:
    delete:
      operationId: revokeManagedInvite
      tags: [Rooms]
      summary: Permanently revoke an invitation as the room owner
      parameters:
        - $ref: '#/components/parameters/RoomSlug'
        - $ref: '#/components/parameters/InviteToken'
      responses:
        '204': {description: Invitation revoked}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
  /rooms/{slug}/legacy-invite:
    delete:
      operationId: disableLegacyRoomInvite
      tags: [Rooms]
      summary: Permanently disable new joins through the room-code link
      parameters:
        - $ref: '#/components/parameters/RoomSlug'
      responses:
        '204': {description: Legacy link disabled}
        '403': {$ref: '#/components/responses/Forbidden'}
  /rooms/{slug}:
    parameters:
      - $ref: '#/components/parameters/RoomSlug'
    get:
      operationId: getRoom
      tags: [Rooms]
      summary: Get a room as a member
      responses:
        '200':
          description: Room details
          content:
            application/json:
              schema:
                type: object
                required: [room]
                properties:
                  room: {$ref: '#/components/schemas/Room'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '410': {$ref: '#/components/responses/Gone'}
    patch:
      operationId: updateRoom
      tags: [Rooms]
      summary: Update room settings as the owner
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/UpdateRoomRequest'}
      responses:
        '200':
          description: Updated room
          content:
            application/json:
              schema:
                type: object
                required: [room]
                properties:
                  room: {$ref: '#/components/schemas/Room'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '410': {$ref: '#/components/responses/Gone'}
        '422': {$ref: '#/components/responses/ValidationError'}
    delete:
      operationId: deleteRoom
      tags: [Rooms]
      summary: Schedule immediate room deletion as the owner
      responses:
        '202': {description: Room closed and scheduled for background deletion}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
  /rooms/{slug}/members:
    get:
      operationId: listRoomMembers
      tags: [Rooms]
      summary: List room participants as the owner
      parameters:
        - $ref: '#/components/parameters/RoomSlug'
      responses:
        '200':
          description: Room owner and joined participants
          content:
            application/json:
              schema:
                type: object
                required: [members, blocked_members]
                properties:
                  members:
                    type: array
                    items: {$ref: '#/components/schemas/RoomMember'}
                  blocked_members:
                    type: array
                    items: {$ref: '#/components/schemas/BlockedRoomMember'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '410': {$ref: '#/components/responses/Gone'}
  /rooms/{slug}/members/{identityID}:
    delete:
      operationId: removeRoomMember
      tags: [Rooms]
      summary: Remove and block a participant as the owner
      parameters:
        - $ref: '#/components/parameters/RoomSlug'
        - name: identityID
          in: path
          required: true
          schema: {type: string, format: uuid}
      responses:
        '204': {description: Participant removed and blocked}
        '400': {$ref: '#/components/responses/BadRequest'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
  /rooms/{slug}/blocked-members/{identityID}:
    delete:
      operationId: unblockRoomMember
      tags: [Rooms]
      summary: Allow a blocked participant to join again
      parameters:
        - $ref: '#/components/parameters/RoomSlug'
        - name: identityID
          in: path
          required: true
          schema: {type: string, format: uuid}
      responses:
        '204': {description: Participant unblocked}
        '400': {$ref: '#/components/responses/BadRequest'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
  /rooms/{slug}/ownership:
    post:
      operationId: transferRoomOwnership
      tags: [Rooms]
      summary: Transfer room ownership to an active participant
      parameters:
        - $ref: '#/components/parameters/RoomSlug'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [identity_id]
              properties:
                identity_id: {type: string, format: uuid}
      responses:
        '200':
          description: Ownership transferred; the caller is now a member
          content:
            application/json:
              schema:
                type: object
                required: [room]
                properties:
                  room: {$ref: '#/components/schemas/Room'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
  /rooms/{slug}/activity:
    get:
      operationId: listRoomActivity
      tags: [Rooms]
      summary: List the latest room activity as the owner
      parameters:
        - $ref: '#/components/parameters/RoomSlug'
      responses:
        '200':
          description: Up to 100 latest events in reverse chronological order
          content:
            application/json:
              schema:
                type: object
                required: [events]
                properties:
                  events:
                    type: array
                    items: {$ref: '#/components/schemas/RoomActivityEvent'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
  /rooms/{slug}/recap:
    get:
      operationId: getRoomRecap
      tags: [Rooms]
      summary: Get full-room event totals as an active member
      parameters:
        - $ref: '#/components/parameters/RoomSlug'
      responses:
        '200':
          description: Event recap totals
          content:
            application/json:
              schema: {$ref: '#/components/schemas/RoomRecap'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
  /rooms/{slug}/highlights:
    get:
      operationId: getRoomHighlights
      tags: [Media]
      summary: Get up to twelve server-ranked highlights across the room
      parameters:
        - $ref: '#/components/parameters/RoomSlug'
      responses:
        '200':
          description: Gallery items ranked by favorites and recency
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    maxItems: 12
                    items: {$ref: '#/components/schemas/GalleryItem'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
  /rooms/{slug}/uploads:
    post:
      operationId: initiateUpload
      tags: [Uploads]
      summary: Reserve capacity and initialize a single or multipart upload
      parameters:
        - $ref: '#/components/parameters/RoomSlug'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/InitiateUploadRequest'}
      responses:
        '201':
          description: Upload initialized
          content:
            application/json:
              schema:
                type: object
                required: [upload, media_id]
                properties:
                  upload: {$ref: '#/components/schemas/Upload'}
                  media_id: {type: string, format: uuid}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/ValidationError'}
        '429': {$ref: '#/components/responses/RateLimited'}
  /uploads/{uploadID}/parts:
    post:
      operationId: getMultipartPartURLs
      tags: [Uploads]
      summary: Issue presigned URLs for a batch of multipart parts
      parameters:
        - $ref: '#/components/parameters/UploadID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [part_numbers]
              properties:
                part_numbers:
                  type: array
                  minItems: 1
                  maxItems: 20
                  uniqueItems: true
                  items: {type: integer, minimum: 1}
      responses:
        '200':
          description: Presigned part URLs
          content:
            application/json:
              schema:
                type: object
                required: [parts]
                properties:
                  parts:
                    type: array
                    items: {$ref: '#/components/schemas/PartURL'}
        '410': {$ref: '#/components/responses/Gone'}
        '422': {$ref: '#/components/responses/ValidationError'}
        '429': {$ref: '#/components/responses/RateLimited'}
  /uploads/{uploadID}/complete:
    post:
      operationId: completeUpload
      tags: [Uploads]
      summary: Verify and complete an upload
      parameters:
        - $ref: '#/components/parameters/UploadID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                parts:
                  type: array
                  description: Required for multipart; omit for a single upload
                  items: {$ref: '#/components/schemas/CompletedPart'}
      responses:
        '200':
          description: Media is ready and thumbnail processing is queued
          content:
            application/json:
              schema:
                type: object
                required: [media]
                properties:
                  media:
                    type: object
                    required: [id, status]
                    properties:
                      id: {type: string, format: uuid}
                      status: {type: string, const: ready}
        '409': {$ref: '#/components/responses/Conflict'}
        '410': {$ref: '#/components/responses/Gone'}
        '422': {$ref: '#/components/responses/ValidationError'}
  /uploads/{uploadID}:
    delete:
      operationId: abortUpload
      tags: [Uploads]
      summary: Abort an upload and release its room reservation
      parameters:
        - $ref: '#/components/parameters/UploadID'
      responses:
        '204': {description: Upload aborted or already inactive}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
  /rooms/{slug}/media:
    get:
      operationId: getRoomGallery
      tags: [Media]
      summary: Get a cursor-paginated gallery
      parameters:
        - $ref: '#/components/parameters/RoomSlug'
        - name: limit
          in: query
          schema: {type: integer, minimum: 1, maximum: 100, default: 50}
        - name: cursor
          in: query
          schema: {type: string}
      responses:
        '200':
          description: Gallery page
          content:
            application/json:
              schema: {$ref: '#/components/schemas/GalleryPage'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '422': {$ref: '#/components/responses/ValidationError'}
  /rooms/{slug}/events:
    get:
      operationId: streamRoomEvents
      tags: [Rooms]
      summary: Stream durable room changes using Server-Sent Events
      description: |
        Sends an initial `ready` event followed by `room` events. Clients reconnect
        with `Last-Event-ID`; heartbeat comments keep idle proxies from closing the
        connection. The stream ends when room access is revoked or expires.
      parameters:
        - $ref: '#/components/parameters/RoomSlug'
        - name: Last-Event-ID
          in: header
          required: false
          schema: {type: integer, format: int64, minimum: 0}
      responses:
        '200':
          description: SSE stream of room events
          headers:
            Cache-Control:
              schema: {type: string, const: 'no-cache, no-transform'}
          content:
            text/event-stream:
              schema:
                type: string
                description: SSE frames whose room-event data matches RoomRealtimeEvent
        '403': {$ref: '#/components/responses/Forbidden'}
        '422': {$ref: '#/components/responses/ValidationError'}
  /media/{mediaID}/download-url:
    post:
      operationId: getMediaDownloadURL
      tags: [Media]
      summary: Issue a presigned original download URL
      parameters:
        - $ref: '#/components/parameters/MediaID'
      responses:
        '200':
          description: Short-lived download URL
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Download'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
...
  /archives/{archiveID}:
    get:
      operationId: getRoomArchive
      tags: [Media]
      summary: Get ZIP generation status and progress
      parameters:
        - $ref: '#/components/parameters/ArchiveID'
      responses:
        '200':
          description: Archive status
          content:
            application/json:
              schema: {$ref: '#/components/schemas/RoomArchiveResponse'}
        '404': {$ref: '#/components/responses/NotFound'}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
...
    ValidationError:
      description: Input validation failed
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
    RateLimited:
      description: Too many requests
      headers:
        Retry-After: {schema: {type: integer}, description: Suggested wait in seconds}
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
  schemas:
...
    AnonymousAuthRequest:
      type: object
      additionalProperties: false
      required: [display_name, client_type]
      properties:
        display_name: {type: string, minLength: 1, maxLength: 80}
        client_type: {type: string, enum: [web, ios, android]}
    TelegramAuthRequest:
      type: object
      additionalProperties: false
      required: [init_data]
      properties:
        init_data:
          type: string
          minLength: 1
          description: Raw, unmodified value from Telegram.WebApp.initData
...
    TokenResponse:
      type: object
      required: [identity, access_token, access_token_expires_in, access_token_expires_at, refresh_token, refresh_token_expires_at]
      properties:
        identity: {$ref: '#/components/schemas/Identity'}
        access_token: {type: string}
        access_token_expires_in: {type: integer, minimum: 0}
        access_token_expires_at: {type: string, format: date-time}
        refresh_token: {type: string}
        refresh_token_expires_at: {type: string, format: date-time}
...
security:
  - bearerAuth: []